added documentation; renamed module file
